Move non-application command stuff out of package
