+++ BotCommands-commands/text/src/main/kotlin/io/github/freya022/botcommands/internal/commands/text/resolvers/users/TextCommandMemberResolverFactory.kt
@@ -8,7 +8,6 @@ import io.github.freya022.botcommands.api.parameters.ResolverRequest
 import io.github.freya022.botcommands.api.parameters.TypedParameterResolverFactory
 import io.github.freya022.botcommands.api.parameters.resolvers.IParameterResolver
 import io.github.freya022.botcommands.api.parameters.resolvers.TextParameterResolver
-import io.github.freya022.botcommands.internal.commands.application.checkGuildOnly
 import net.dv8tion.jda.api.entities.Member
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent
 import java.util.regex.Pattern
@@ -24,7 +23,6 @@ internal class TextCommandMemberResolverFactory(
     private val adapter = Adapter()
 
     override fun get(request: ResolverRequest): IParameterResolver<*> {
-        request.checkGuildOnly(Member::class)
         return adapter
     }
